backend/requirements.txt
django==4.2.3
djangorestframework==3.14.0
gunicorn==20.1.0
django-cors-headers==4.2.0
mysqlclient==2.2.0